backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-google-genai
langchain-core
requests
langchain-openai